Tests/LibWeb/Text/expected/abspos-relayout-inset-stretch.txt

boundary=0,0,320,180
child=0,0,100,70
partial-layouts=1 full-layouts=0
PASS
